refactor: Components validators separation
